chore: generate search indexes and tags at build time
