+++ public/tags/java-monitoring/feed.xml
@@ -1,19 +1,19 @@
 
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
     <channel>
-      <title>OpenTelemetry-Native Traces, Metrics, and Logs in a single pane</title>
-      <link>https://www.signoz.io/blog</link>
-      <description>A single tool for all your observability needs - APM, logs, metrics, exceptions, alerts, and dashboards powered by a powerful query builder.</description>
+      <title>SigNoz</title>
+      <link>https://signoz.io/blog</link>
+      <description>SigNoz is an open-source observability tool to help you find issues in your deployed applications &amp; solve them quickly. It provides logs, metrics, and traces under a single pane of glass. Built with ClickHouse as datastore, it is designed to handle enterprise scale.</description>
       <language>en-us</language>
       <managingEditor>address@yoursite.com (SigNoz Inc)</managingEditor>
       <webMaster>address@yoursite.com (SigNoz Inc)</webMaster>
       <lastBuildDate>Thu, 12 Oct 2023 00:00:00 GMT</lastBuildDate>
-      <atom:link href="https://www.signoz.io/tags/java-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
+      <atom:link href="https://signoz.io/tags/java-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
       
   <item>
-    <guid>https://www.signoz.io/blog/2022-03-10-everything-you-need-to-know-about-opentelemetry-java-agent</guid>
+    <guid>https://signoz.io/blog/java-agent</guid>
     <title>OpenTelemetry Java Agent - Implement Observability With Zero Code Changes</title>
-    <link>https://www.signoz.io/blog/2022-03-10-everything-you-need-to-know-about-opentelemetry-java-agent</link>
+    <link>https://signoz.io/blog/java-agent</link>
     undefined
     <pubDate>Thu, 12 Oct 2023 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
@@ -21,9 +21,9 @@
   </item>
 
   <item>
-    <guid>https://www.signoz.io/blog/java-auto-instrumentation</guid>
+    <guid>https://signoz.io/blog/java-auto-instrumentation</guid>
     <title>OpenTelemetry Java auto-instrumentation - Everything you need to know</title>
-    <link>https://www.signoz.io/blog/java-auto-instrumentation</link>
+    <link>https://signoz.io/blog/java-auto-instrumentation</link>
     undefined
     <pubDate>Tue, 17 Aug 2021 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
@@ -31,19 +31,9 @@
   </item>
 
   <item>
-    <guid>https://www.signoz.io/blog/2021-08-17-opentelemetry-java-auto-instrumentation</guid>
-    <title>OpenTelemetry Java auto-instrumentation - Everything you need to know</title>
-    <link>https://www.signoz.io/blog/2021-08-17-opentelemetry-java-auto-instrumentation</link>
-    undefined
-    <pubDate>Tue, 17 Aug 2021 00:00:00 GMT</pubDate>
-    <author>address@yoursite.com (SigNoz Inc)</author>
-    <category>java-monitoring</category>
-  </item>
-
-  <item>
-    <guid>https://www.signoz.io/blog/tomcat</guid>
+    <guid>https://signoz.io/blog/tomcat</guid>
     <title>Monitor Tomcat Java application with OpenTelemetry and SigNoz</title>
-    <link>https://www.signoz.io/blog/tomcat</link>
+    <link>https://signoz.io/blog/tomcat</link>
     undefined
     <pubDate>Wed, 17 Aug 2022 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
