+++ public/tags/javascript-monitoring/feed.xml
@@ -1,19 +1,19 @@
 
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
     <channel>
-      <title>OpenTelemetry-Native Traces, Metrics, and Logs in a single pane</title>
-      <link>https://www.signoz.io/blog</link>
-      <description>A single tool for all your observability needs - APM, logs, metrics, exceptions, alerts, and dashboards powered by a powerful query builder.</description>
+      <title>SigNoz</title>
+      <link>https://signoz.io/blog</link>
+      <description>SigNoz is an open-source observability tool to help you find issues in your deployed applications &amp; solve them quickly. It provides logs, metrics, and traces under a single pane of glass. Built with ClickHouse as datastore, it is designed to handle enterprise scale.</description>
       <language>en-us</language>
       <managingEditor>address@yoursite.com (SigNoz Inc)</managingEditor>
       <webMaster>address@yoursite.com (SigNoz Inc)</webMaster>
       <lastBuildDate>Mon, 20 Nov 2023 00:00:00 GMT</lastBuildDate>
-      <atom:link href="https://www.signoz.io/tags/javascript-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
+      <atom:link href="https://signoz.io/tags/javascript-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
       
   <item>
-    <guid>https://www.signoz.io/blog/nodejs</guid>
+    <guid>https://signoz.io/blog/nodejs</guid>
     <title>OpenTelemetry Setup in a Nodejs Application</title>
-    <link>https://www.signoz.io/blog/nodejs</link>
+    <link>https://signoz.io/blog/nodejs</link>
     undefined
     <pubDate>Mon, 20 Nov 2023 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
