+++ public/tags/opentelemetry-tutorials/feed.xml
@@ -1,21 +1,121 @@
 
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
     <channel>
-      <title>OpenTelemetry-Native Traces, Metrics, and Logs in a single pane</title>
-      <link>https://www.signoz.io/blog</link>
-      <description>A single tool for all your observability needs - APM, logs, metrics, exceptions, alerts, and dashboards powered by a powerful query builder.</description>
+      <title>SigNoz</title>
+      <link>https://signoz.io/blog</link>
+      <description>SigNoz is an open-source observability tool to help you find issues in your deployed applications &amp; solve them quickly. It provides logs, metrics, and traces under a single pane of glass. Built with ClickHouse as datastore, it is designed to handle enterprise scale.</description>
       <language>en-us</language>
       <managingEditor>address@yoursite.com (SigNoz Inc)</managingEditor>
       <webMaster>address@yoursite.com (SigNoz Inc)</webMaster>
-      <lastBuildDate>Fri, 08 Sep 2023 00:00:00 GMT</lastBuildDate>
-      <atom:link href="https://www.signoz.io/tags/opentelemetry-tutorials/feed.xml" rel="self" type="application/rss+xml"/>
+      <lastBuildDate>Thu, 06 Jun 2024 00:00:00 GMT</lastBuildDate>
+      <atom:link href="https://signoz.io/tags/opentelemetry-tutorials/feed.xml" rel="self" type="application/rss+xml"/>
       
   <item>
-    <guid>https://www.signoz.io/blog/2022-05-20-opentelemetry-go</guid>
+    <guid>https://signoz.io/blog/add-manual-span-to-traces-nodejs</guid>
+    <title>Manual Instrumentation for Traces - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/add-manual-span-to-traces-nodejs</link>
+    undefined
+    <pubDate>Thu, 06 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/autoinstrumented-tracing-nodejs</guid>
+    <title>Autoinstrumentation for Traces - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/autoinstrumented-tracing-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/collector-nodejs</guid>
+    <title>Setting up the Otel Collector - OpenTelemetry NodeJSS</title>
+    <link>https://signoz.io/blog/collector-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/correlating-traces-logs-metrics-nodejs</guid>
+    <title>Correlating Traces, Logs, and Metrics - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/correlating-traces-logs-metrics-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/custom-metrics-nodejs</guid>
+    <title>Setting up Custom Metrics - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/custom-metrics-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/go</guid>
     <title>Complete guide to implementing OpenTelemetry in Go applications</title>
-    <link>https://www.signoz.io/blog/2022-05-20-opentelemetry-go</link>
+    <link>https://signoz.io/blog/go</link>
     undefined
     <pubDate>Fri, 08 Sep 2023 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/logging-nodejs</guid>
+    <title>Sending Logs to SigNoz - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/logging-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/metrics-nodejs</guid>
+    <title>Exploring Metrics created via Traces in SigNoz - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/metrics-nodejs</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/nodejs-clone-application</guid>
+    <title>Understanding and Cloning the Sample Application - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/nodejs-clone-application</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/nodejs-docker-setup</guid>
+    <title>Setting Up Docker and Your Local Environment - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/nodejs-docker-setup</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
+    <author>address@yoursite.com (SigNoz Inc)</author>
+    <category>opentelemetry-tutorials</category>
+  </item>
+
+  <item>
+    <guid>https://signoz.io/blog/nodejs-tutorial-overview</guid>
+    <title>Overview - Implementing OpenTelemetry in NodeJS with SigNoz - OpenTelemetry NodeJS</title>
+    <link>https://signoz.io/blog/nodejs-tutorial-overview</link>
+    undefined
+    <pubDate>Wed, 05 Jun 2024 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
     <category>opentelemetry-tutorials</category>
   </item>
