+++ public/tags/python-monitoring/feed.xml
@@ -1,19 +1,19 @@
 
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
     <channel>
-      <title>OpenTelemetry-Native Traces, Metrics, and Logs in a single pane</title>
-      <link>https://www.signoz.io/blog</link>
-      <description>A single tool for all your observability needs - APM, logs, metrics, exceptions, alerts, and dashboards powered by a powerful query builder.</description>
+      <title>SigNoz</title>
+      <link>https://signoz.io/blog</link>
+      <description>SigNoz is an open-source observability tool to help you find issues in your deployed applications &amp; solve them quickly. It provides logs, metrics, and traces under a single pane of glass. Built with ClickHouse as datastore, it is designed to handle enterprise scale.</description>
       <language>en-us</language>
       <managingEditor>address@yoursite.com (SigNoz Inc)</managingEditor>
       <webMaster>address@yoursite.com (SigNoz Inc)</webMaster>
       <lastBuildDate>Tue, 13 Feb 2024 00:00:00 GMT</lastBuildDate>
-      <atom:link href="https://www.signoz.io/tags/python-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
+      <atom:link href="https://signoz.io/tags/python-monitoring/feed.xml" rel="self" type="application/rss+xml"/>
       
   <item>
-    <guid>https://www.signoz.io/blog/opentelemetry/python</guid>
+    <guid>https://signoz.io/blog/python</guid>
     <title>Monitor your Python application with OpenTelemetry and SigNoz</title>
-    <link>https://www.signoz.io/blog/opentelemetry/python</link>
+    <link>https://signoz.io/blog/python</link>
     undefined
     <pubDate>Tue, 13 Feb 2024 00:00:00 GMT</pubDate>
     <author>address@yoursite.com (SigNoz Inc)</author>
